Create LICENSE
From both this publicly published repository and the fact that you've extensively explained how to implement your code, it seems to me that you want this code to be in the public domain. Including a license would help very much with that. I've chosen the MIT license, but feel free to include any [other license](https://help.github.com/articles/licensing-a-repository/) if you so please -- if you intend to include one, that is.